refactor(core): 重构代码结构和清理冗余代码
- 重新排列 __init__.py 中的导出顺序并按功能分组
- 移除 services 模块中的冗余 else 语句
- 删除 annotation_items.py 中未使用的 QTransform 和 QGraphicsPixmapItem 导入
- 格式化 annotation_items.py 中的 QPen 设置代码以提高可读性
- 简化 autostart.py 中的平台检测逻辑
- 移除 base_tab.py 中抽象方法的空 pass 语句
- 更新 base_worker.py 中的泛型语法并移除未使用的 Generic 导入
- 修复 batch_queue_manager.py 中的数值转换问题
- 优化 clipboard_controller.py 中的 CF_HTML 生成逻辑
- 添加 TYPE_CHECKING 导入以解决循环依赖问题
- 移除 layout.json 中的几何配置差异
- 重构 log_service.py 中的信号发射器实现
- 格式化 magnifier_overlay.py 中的十字线绘制坐标
- 简化 main.py 中的托盘菜单连接逻辑
- 更新 main_window.py 中的类型检查导入位置
- 修复 markdown_converter.py 中的条件判断逻辑
- 优化模型缓存管理器中的状态检查流程
- 更新 OCR 服务中的 zip 调用以使用 strict 参数
- 清理 OCR 服务基础类中的抽象方法空实现
- 优化便携版 OCR 服务的状态返回逻辑

+++ src/vibeocr/workers/batch_queue_manager.py
@@ -242,7 +242,7 @@ class BatchQueueManager:
         if self._memory_monitor.is_available():
             # 假设平均图像尺寸约为 1920x1080 (约 2M 像素)
             # 实际应该解码图像获取真实尺寸，这里用简化的估算
-            estimated_pixels = avg_size / 3 * 2  # 粗略估计
+            estimated_pixels = int(avg_size / 3 * 2)  # 粗略估计
             batch_size = self._memory_monitor.estimate_batch_size(estimated_pixels)
         else:
             batch_size = self.max_batch_size
